# public/openapi.yaml
openapi: 3.0.3
info:
  title: Daftra E-commerce API
  description: |
    RESTful API for Daftra E-commerce platform providing authentication, user management, and comprehensive catalog browsing endpoints.
    
    ## Authentication
    This API uses Bearer token authentication with Laravel Sanctum. Include the token in the Authorization header:
    `Authorization: Bearer {token}`
    
    ## Security Features
    - Timing attack prevention on login endpoints
    - Secure password hashing with bcrypt
    - Rate limiting on authentication endpoints
    
    ## Performance Features
    - Intelligent caching on catalog endpoints (1-hour cache)
    - Optimized database queries with proper indexing
    - Fast response times with minimal database load
    - Efficient pagination for large datasets
    
    ## Catalog Features
    - Advanced product filtering by category, price range, and search
    - Laravel-standard pagination with metadata
    - Products ordered by price for consistent browsing
    - Real-time stock information
    - Thumbnail image support
    
    ## Response Format
    All responses follow a consistent JSON format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: Daftra API Support
    email: support@daftra.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.daftra.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /checkout:
    post:
      tags:
        - Orders
      summary: Checkout (Create Order)
      description: |
        Create a new order for the authenticated user. Requires Bearer token.
      operationId: checkout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - product_id
                      - quantity
                    properties:
                      product_id:
                        type: integer
                        example: 1
                      quantity:
                        type: integer
                        example: 2
      responses:
        '200':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Server error (invalid product, etc)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders:
    get:
      tags:
        - Orders
      summary: Get Orders List
      description: |
        Retrieve a paginated list of the authenticated user's orders. Requires Bearer token.
      operationId: getOrdersList
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPaginatedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get Order Details
      description: |
        Retrieve details for a specific order belonging to the authenticated user. Requires Bearer token.
      operationId: getOrderDetails
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
          description: ID of the order to retrieve
      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found or does not belong to user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Laravel Sanctum Bearer token authentication.
        
        **How to obtain a token:**
        1. Use the `/login` endpoint with valid credentials
        2. Include the returned token in the `Authorization` header as `Bearer {token}`
        
        **Token format:**
        `{token_id}|{plain_text_token}`
        
        **Example:**
        `Authorization: Bearer 1|dy3hZHSBvfqAhhHm48ar8g2mfnFAb1K1x2a3fSxL884c2a58`
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Invalid credentials"
      example:
        message: "Invalid credentials"
    ValidationErrorResponse:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          description: General validation error message
          example: "The given data was invalid."
        errors:
          type: object
          description: Field-specific validation errors
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - "The email field is required."
            password:
              - "The password field is required."
      example:
        message: "The given data was invalid."
        errors:
          email:
            - "The email field is required."
          password:
            - "The password field is required."
...
    Order:
      type: object
      required:
        - id
        - subtotal
        - shipping
        - tax
        - total
        - items
      properties:
        id:
          type: integer
          example: 1
        subtotal:
          type: integer
          description: Subtotal in cents
          example: 2000
        shipping:
          type: integer
          description: Shipping in cents
          example: 1500
        tax:
          type: number
          description: Tax amount in cents
          example: 350
        total:
          type: number
          description: Total amount in cents
          example: 3850
        order_number:
          type: integer
          description: Sequential order number for the user
          example: 2
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderInList:
      type: object
      required:
        - id
        - subtotal
        - shipping
        - tax
        - total
        - items
      properties:
        id:
          type: integer
          example: 1
        subtotal:
          type: integer
          description: Subtotal in cents
          example: 2000
        shipping:
          type: integer
          description: Shipping in cents
          example: 1500
        tax:
          type: number
          description: Tax amount in cents
          example: 350
        total:
          type: number
          description: Total amount in cents
          example: 3850
        order_number:
          type: integer
          description: Sequential order number for the user
          example: 2
    OrderItem:
      type: object
      required:
        - product_id
        - quantity
        - title
        - image_url
        - price
        - total
      properties:
        product_id:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        title:
          type: string
          example: "iPhone 14 Pro"
        image_url:
          type: string
          format: uri
          example: "https://example.com/thumbnails/iphone14.jpg"
        price:
          type: integer
          description: Price per item in cents
          example: 1000
        total:
          type: integer
          description: Total for this item (price * quantity)
          example: 2000
    OrderPaginatedResponse:
      type: object
      required:
        - data
        - links
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/OrderInList'
        links:
          type: object
          required:
            - first
            - last
            - prev
            - next
          properties:
            first:
              type: string
              format: uri
            last:
              type: string
              format: uri
            prev:
              type: string
              format: uri
              nullable: true
            next:
              type: string
              format: uri
              nullable: true
        meta:
          type: object
          required:
            - current_page
            - from
            - last_page
            - per_page
            - to
            - total
          properties:
            current_page:
              type: integer
            from:
              type: integer
              nullable: true
            last_page:
              type: integer
            per_page:
              type: integer
            to:
              type: integer
              nullable: true
            total:
              type: integer
...